+++ src/Events/TestJobFinished.php
@@ -1,0 +1,18 @@
+<?php
+declare(strict_types=1);
+
+namespace MyTester\Events;
+
+use MyTester\Job;
+
+/**
+ * Triggers when execution of a test job finished
+ *
+ * @author Jakub Konečný
+ */
+final readonly class TestJobFinished
+{
+    public function __construct(public Job $job)
+    {
+    }
+}